requirements.txt
# Core LangChain (no torch dependencies)
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-text-splitters==0.0.1
langchain-openai==0.1.7

# Vector store with built-in embeddings (NO PYTORCH!)
chromadb==0.4.22

# Data processing
pandas==2.1.4
numpy==1.26.3

# LLM client
openai==1.7.2
requests==2.31.0

# Forecasting (NO PROPHET - uses pure statsmodels)
statsmodels==0.14.1
scikit-learn==1.3.2
prophet==1.1.5

# Utilities
pyyaml==6.0.1
colorama==0.4.6
prompt-toolkit==3.0.43
python-dotenv==1.0.0